Added Clippy warnings with Git hook (#15)
- Added warnings for cargo, pedantic, and nursery warnings. Suppressed with explanatory comments or fixed resulting warnings.

- Added Git hooks for formatting, linting, and tests via cargo-husky

- Removed unnecessary Clippy statements

+++ src/components/enemy.rs
@@ -39,15 +39,15 @@ pub fn initialize_enemies(world: &mut World) {
     use rand::distributions::{Distribution, Uniform};
     let mut rng = rand::thread_rng();
     let enemy_count = Uniform::new(1, 20);
-    let direction = Uniform::new(-1, 1);
+    let direction = Uniform::new(-1.0, 1.0);
     let velocity = Uniform::new(f32::EPSILON, 50.0);
     let enemy_x = Uniform::new(HITCIRCLE_RADIUS, ARENA_WIDTH - HITCIRCLE_RADIUS);
     let enemy_y = Uniform::new(HITCIRCLE_RADIUS, ARENA_HEIGHT - HITCIRCLE_RADIUS);
     for _ in 1..=enemy_count.sample(&mut rng) {
         spawn_enemy(
             world,
-            direction.sample(&mut rng) as f32 * enemy_x.sample(&mut rng),
-            direction.sample(&mut rng) as f32 * enemy_y.sample(&mut rng),
+            direction.sample(&mut rng) * enemy_x.sample(&mut rng),
+            direction.sample(&mut rng) * enemy_y.sample(&mut rng),
             velocity.sample(&mut rng),
             velocity.sample(&mut rng),
         );
